chore: used shiki plugin as a afterSetup plugin
